buildbot: build-slave-package: Fix build failure of missing XU4 DTS on stable v4.1
Linux kernel stable v4.1 does not have Odroid HC1 and Odroid XU4,
thus it failed:

    No Odroid HC1 DTS, using regular XU4
    cp: cannot stat 'out//arch/arm/boot/dts/exynos5422-odroidxu4.dtb': No such file or directory

+++ buildbot/build-slave-package.sh
@@ -29,9 +29,14 @@ if [ ! -f "arch/arm/boot/dts/exynos5422-odroidxu3-lite.dts" ]; then
 	cp ${OUT_DIR}arch/arm/boot/dts/exynos5422-odroidxu3.dtb ${OUT_DIR}arch/arm/boot/dts/exynos5422-odroidxu3-lite.dtb
 fi
 if [ ! -f "arch/arm/boot/dts/exynos5422-odroidhc1.dts" ]; then
-	# If there is no Odroid HC1 DTS, then duplicate the XU4 DTB
-	echo "No Odroid HC1 DTS, using regular XU4"
-	cp ${OUT_DIR}arch/arm/boot/dts/exynos5422-odroidxu4.dtb ${OUT_DIR}arch/arm/boot/dts/exynos5422-odroidhc1.dtb
+	# All compiled kernels have SMDK5410 and XU3 DTS but for example v4.1 does not have XU4 DTS
+	if [ -f "arch/arm/boot/dts/exynos5422-odroidxu4.dts" ]; then
+		# If there is no Odroid HC1 DTS, then duplicate the XU4 DTB
+		echo "No Odroid HC1 DTS, using regular XU4"
+		cp ${OUT_DIR}arch/arm/boot/dts/exynos5422-odroidxu4.dtb ${OUT_DIR}arch/arm/boot/dts/exynos5422-odroidhc1.dtb
+	else
+		echo "No Odroid HC1 DTS and no XU4 DTS."
+	fi
 fi
 
 # Remove old modules-out
